Add to_sys_name() to replace the win32 tests
A holdover from DOS 8.3 filenames, files on Windows cannot end with a
dot.  We append a - to such names on Windows platforms in all
operations, which should solve the problem, but we'd just duplicated
that code about a dozen times.  No need, do it once and we can add
whatever filesystem rules for the host system we need to in one spot.