Some diff reduction and renamed variables
A few of my local copies of cppo have some/most of the code reformatted
in a more "pythonic" coding style.  I still use hard tabs for
indentation because even us diehard console-using developers have
editors that can have whatever tabstop we want on a per-file basis, and
we have editorconfig.  It's 2017, get with the times, even for a program
made for accessing files for a 1977 computer!  ;)  No functional changes
here, save for an if statement processing extensions replaces multiple
conditionals with an if x in tuple construct.