Protect main program with __main__ test
Originally cppo was written as a shell script and was never intended to
be a library of functions to be used by anyone else.  Single-file Python
modules are often written to be run as standalone programs either to do
what they do from the command line or for testing purposes.

Theoretically you do this if your code provides useful stuff for other
programs to use, and it's hard to argue that cppo does that yet, but it
is intended to do so in the future.  Let's start working toward that.