Document two more test disks
We're using the latest community GSOS release as a convenient example of a
large ProDOS volume with lots of directory structure, forks, and at least a few
files that can be identified by eyeball as having been extracted correctly.

Then I went looking for something DOS 3.3 with T type files on the disk to
verify those were being extracted correctly.  Randomly stumbled across The
Correspondent 4.4 which has them, but also had some surprises for us.  It
crashed cppo!  A combination of removing bashbyter functions and ensuring a
variable gets initialized caused cppo to be able to cat and dump it.  We do
nothing special about filenames made up entirely of control characters,
however, so they'll print incorrectly and extract as non-printable characters
in your filesystem.  It's legal (if crazy) on most UNIX filesystems.  I have
_no idea_ if or how to handle these things on macOS or Windows.