Logging, unStudlyCapping, main()
Replaced some of the commented out print lines with calls to a logger object.
These are currently always turned on, which we don't want, but they don't hurt
anything and are short.

Oh, and new logging system!  The setup is a bit more verbose than it could be
because logging predates str.format and uses the str % tuple syntax that now
exists mostly to avoid breaking older code.  You can override this, and I did.
It's not done yet because we might want to actually make some of the existing
print calls into log.info's.  Y'know, just as soon as I set that up so ONLY
logging.INFO goes to stdout, unadorned, and everything higher than that goes to
stderr, depending on your logging level, with pretty formatting.

Yeah, logging can do all of that and chew bubblegum at the same time, I just
haven't set it up yet because I want to do it right.

A little more unStudlyCapping of things.  I'm going to have to start actually
creating classes soon which is going to bring back the capitals, but I've been
working to get rid of them so that it becomes less confusing when we get there.
I dunno if it's helped any.

I also added a few comments about our imports and checked that we actually used
everything we imported.  No, we don't.  But we maybe should switch what we are
using for what we aren't at some point?