Use struct.{un,}pack_into instead of int methods
The methods int.to_bytes() and int.from_bytes() are ... both messy and ugly,
contrary to the Zen of Python.  Right now, we're just writing out single ints
and words, but eventually we'll be reading and writing whole structures, so it
makes sense to begin moving the codebase in that direction.

I've written only the functions I either use or see a use for immediately, save
for pack_u32be() which I wrote for date conversions, and then immediately
realized we don't want to use for that purpose yet.  We can remove it if we
don't ultimately need it.