First move toward Disk class w/ a2mg data
Finally!  The Disk class doesn't actually serve as much more than a slightly
improved Globals class at the moment holding every splitting of the source path
and filename that we use in legacy code, as well as a copy of the disk image
itself that gets used long enough to read the a2mg header.

The idea I have here is to begin building the module-based code in parallel.
Then I'll just modify the linear code to compare doing it the old way to doing
it the new.  That'll let me verify that the new code does what the old should.

When it's all done, we can just modify main to use the new modular code and
look at splitting the modular code into a package with cppo as a runner.  At
that point the code should begin being able to do things cppo cannot.  We could
continue to extend cppo at that point, but my inclination is to maintain the
cppo runner as a compatibility layer and begin building a more modern image
tool.  Essentially to begin building the CiderPress for Linux or the Java-free
AppleCommander.