Alternate 2mg implementation
Rwrite of the 2mg parsing code which saves the comment and creator blocks in
memory by putting the code into the Disk class.

Doesn't yet attempt to parse out 2mg image format, and the question is open as
to whether or not I'll try to follow cppo's historical method of handling this
(chop the header and go), AppleCommander's method (consider 2mg its own unique
image format which happens to contain one of the others) or CiderPress's (use
one of several possible wrappers independent of the image orders.  CP's is
probably the most flexible.

So basically this code is probably far from final, but it works for what it
does so far.