Use textwrap to dedent multi-line strings
This is kind of an expensive thing to do unconditionally, but it lets us make
multi-line strings fit into the code with less ugliness.  Basically, if you're
four levels in, you can do something like this:

                log.warn("""\
                There was a problem.
                It was probably wasn't fatal because this
                is only a warning, but it is enough to have
                a multiline string.
                """)

This will print without the indentation.  It's not quite as clean as how
docutils handles docstrings (allowing the first line to be unindented, hence
the line-continuation), but it's still an improvement.  If you can improve upon
this, please feel free to PR it!