Add license, Copyright notices, history doc
The history document is kind of a mishmash of explanation about what decisions
have lead to what this project is trying to do and why this rather than other
things, such as improving AppleCommander.  (Ohh, it has the reason for that
believe me--die in the cash-consuming fire of the Internet's rage, Oracle!)

More importantly, there are Copyright notices and the GNU GPL v2.