Broke diskimg.py out of legacy
I may not have done this 100% "properly"--this is really the first full
application thingy I've ever tried to write in Python.  I learned circular
imports are possible and the error messages are not obvious when you do that.

I've also learned that importing a package doesn't necessarily import the
modules within that package--if you want that, the package needs to import its
submodules into itself.  That was not obvious either.  It does explain why
sometimes you must import subpackages directly and other times importing the
package is all you need.  This is probably obvious to Python developers who
actually work on big projects, but ... I've only ever done single-file scripts
before now so it wasn't obvious to me.

For now, blocksfree is importing legacy.  We don't have enough outside of
legacy yet to make the alternative even remotely useful at this time.
Eventually though the goal is to stop doing that.