Added an abstract buffer class
What this is missing is a flush method.  It's not implemented because:

a. cppo is all we've got so far and it doesn't need write at all
b. I'm not 100% sure how I'm doing files yet
c. I'm hoping the way I ought to implement this will make itself apparent when
the code gets used.