Correct typing and docstring on seqsplit
Type hinting in Python is just that: Hinting.  Which means if you do it wrong,
as long as the syntax is right, Python itself will just let you do it.  Python
is intended to be a duck-typed language, so you're really just documenting you
intent/expectations.  The feature is somewhat new, and its conventions are not
completely understood by certain tools that encounter it.  This apparently
applies to myself as a developer trying to use it. ;)