Mostly style improvements to logging.py
Our changes to the built-in logging module of Python are kind of a hack designed
to be as lightweight as possible way to replace the built-in logging module with
one that operates using newer str.format based string expansion.  It's not
really complete we probably should change that at some point.

Changes include:

 - Docstrings, lots of docstrings
 - Type hinting
 - log is now LOG
 - pylint warnings disabled on things that will not change and are on purpose
 - StyleAdapter.log does not dedent msg anymore unless dedent=True is passed
   which hopefully should make it a little less DWIM.