Hexdump, better __str__ for ByteBuffer
We broke str(ByteBuffer) with the util.py commit, but having that return a
possibly megabytes-long string was stupid anyway.  It now returns a string just
indicating that it's not really a printable thing rather than the still
probably extremely long repr().

Added a hexdump function to call util.hexdump cleanly.  This is really only
here for debugging.  I can imagine a lot of things you could do to account for
possibilities later on, but YAGNI applies (including that if you do need it,
you can add it then.)