Change ByteBuffer reads on one byte
If you read one index from a bytes or bytearray, you get an int, not a single
character bytes object.  Originally I didn't want to mimic that feature because
it's actually somewhat annoying at times.  Realizing it was done that way for a
reason, and not doing it in ByteBuffer is gonna be even more annoying.