Add git files and cppo to .editorconfig
If the use of a four-space tab for files under .git aside from patches (and
commit messages if you use commit.verbose) is annoying to you, feel free to PR
this to make the .editorconfig more specific.

Additionally, cppo was not explicitly specified in editorconfig.  EditorConfig
applies rules based on filename, not filetype.  That doesn't bother me--there is
a vim modeline in the fiel.  It might bother you if you use something else.