Use LOG, move logging setup to cppo
The actual logging configuration belongs in the application, not a module.  In a
module, it's always going to do what the module says to do.  And that's fine for
cppo which is a script that works the way it does, but we have bigger plans for
this code.  It's now in the cppo module.

We've stopped using log from blocksfree.logging, favoring LOG instead, so we
have removed it.