Lint cppo: Docstrings, warning disables
This actually disables all pylint warnings we haven't fixed because, as of now,
we don't intend to fix them.  The arg parsing is suboptimal, but if you can make
sure that existing scripts using cppo don't break by turning this into argparse
code, be my guest.  I didn't see an obvious way to do it any better than what we
have now.  When cppo is rewritten, we'll create a new runner with its own
argparse-based command parser.  Until then, I'm inclined to leave it be.

This is just part of a larger rewrite I started working on, but ... it's not
needed now.