Remove 2img code, use ByteBuffer instead
One thing both AppleCommander and CiderPress do is wrap 2img is effectively
treat 2img files as containers.  That's a good idea.  We need to start using our
buffer code before it begins making sense, but that means this 2img code has to
go as it's kind of just bolted on to the side for now.

Instead we'll put the image we load into a buffer.  How we do that is likely to
change, but this gets us to the point we can start using it.